+++ glue/qt/qtutil.py
@@ -13,10 +13,10 @@ from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui, QtCore
 from glue import core
 from glue.qt import ui, icons
-from glue.qt.decorators import set_cursor
 from glue.qt.mime import LAYERS_MIME_TYPE
 from glue.utils.qt import (QMessageBoxPatched as QMessageBox, mpl_to_qt4_color,
-                           qt4_to_mpl_color, tint_pixmap, GlueItemWidget)
+                           qt4_to_mpl_color, tint_pixmap, GlueItemWidget,
+                           set_cursor)
 
 
 def data_wizard():
