+++ glue/qt/widgets/data_viewer.py
@@ -6,10 +6,10 @@ from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui
 from glue.core.application_base import ViewerBase
 from glue.qt import get_qapp
-from glue.qt.decorators import set_cursor
 from glue.qt.layer_artist_model import QtLayerArtistContainer, LayerArtistView
 from glue.qt.mime import LAYERS_MIME_TYPE, LAYER_MIME_TYPE
 from glue.qt.widgets.glue_mdi_area import GlueMdiSubWindow
+from glue.utils.qt import set_cursor
 
 
 __all__ = ['DataViewer']
