+++ glue/utils/qt/__init__.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, division, print_function
 from .autocomplete_widget import *
 from .qmessagebox_widget import *
 from .colors import *
+from .decorators import *
 from .dialogs import *
 from .helpers import *
 from .mixins import *
